# project_back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Exercices",
        "version": "1.0"
    },
    "paths": {
...
        "/trainings/user/{id}": {
            "get": {
                "tags": [
                    "Trainings"
                ],
                "summary": "Récupère les séances d'un utilisateur avec leurs exercices",
                "operationId": "1a76d9d96661f4e5c2c17ce9574e2058",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des séances de l'utilisateur récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Training"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune séance trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Aucune séance trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/trainings/admin": {
            "get": {
                "tags": [
                    "Trainings"
                ],
                "summary": "Récupère toutes les séances (Admin)",
                "operationId": "6eee762b7df81483573160955f58516d",
                "responses": {
                    "200": {
                        "description": "Liste des séances récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Training"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune séance trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Aucune séance trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/trainings/{id}": {
            "get": {
                "tags": [
                    "Trainings"
                ],
                "summary": "Affiche les détails d'une séance avec ses exercices",
                "operationId": "6da151952408cf2a17856fa48e92dfec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la séance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la séance récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Training"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Séance non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Séance non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Trainings"
                ],
                "summary": "Mettre à jour une séance",
                "operationId": "7d38377b1b09fe961b76b98d481b7b29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la séance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Séance modifiée"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Séance mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Training"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Séance non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Séance non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Trainings"
                ],
                "summary": "Supprimer une séance",
                "operationId": "680286f19ac3b9507144691464db73de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la séance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Séance supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Séance supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Séance non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Séance non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/trainings": {
            "post": {
                "tags": [
                    "Trainings"
                ],
                "summary": "Créer une nouvelle séance",
                "operationId": "ff12987a61e7ecbd56811ad4ee47872b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Séance du matin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Séance créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Training"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Exercice": {
                "title": "Exercice",
                "description": "Un modèle représentant un exercice d'entraînement",
                "properties": {
                    "id": {
                        "description": "ID de l'exercice",
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "description": "Nom de l'exercice",
                        "type": "string",
                        "example": "Pompes"
                    },
                    "description": {
                        "description": "Description de l'exercice",
                        "type": "string",
                        "example": "Exercice de renforcement musculaire"
                    },
                    "series": {
                        "description": "Nombre de séries",
                        "type": "integer",
                        "example": 3
                    },
                    "repetitions": {
                        "description": "Nombre de répétitions",
                        "type": "integer",
                        "example": 12
                    },
                    "training_id": {
                        "description": "ID de la séance associée à cet exercice",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Date de création de l'exercice",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour de l'exercice",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Training": {
                "title": "Training",
                "description": "Un modèle représentant une séance d'entraînement",
                "properties": {
                    "id": {
                        "description": "ID de la séance",
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "description": "Nom de la séance",
                        "type": "string",
                        "example": "Séance de musculation"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur propriétaire de la séance",
                        "type": "integer",
                        "example": 1
                    },
                    "exercices": {
                        "description": "Liste des exercices associés à la séance",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Exercice"
                        }
                    },
                    "created_at": {
                        "description": "Date de création de la séance",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour de la séance",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}